change the nominal from float to long double

+++ model/model_pos_anggaran/cekPos_anggaran.c
@@ -7,11 +7,11 @@ bool cekPos_anggaran(char *data){
       if (cekFile != NULL) {
         char line[100];
         char pos_existing[50];
-        float batas_existing;
+        long double batas_existing;
 
         while (fgets(line, sizeof(line), cekFile)) {
           // format file: namaPos|batasNominal
-          if (sscanf(line, "%49[^|]|%f", pos_existing, &batas_existing) == 2) {
+          if (sscanf(line, "%49[^|]|%Lf", pos_existing, &batas_existing) == 2) {
             if (strcmp(pos_existing, data) == 0) {
               return true;
               break;
