+++ model/model_pos_anggaran/tambah_pos_anggaran_model.c
@@ -2,13 +2,13 @@
 #include <stdbool.h>
 #include <string.h>
 
-void tambah_pos_anggaran_model(char *nama_pos, float nominal){
+void tambah_pos_anggaran_model(char *nama_pos, long double nominal){
     FILE *fp = fopen("pos_anggaran.txt", "a");
       if (fp == NULL) {
         printf("Gagal membuka file untuk menulis!\n");
         return;
       }
-      fprintf(fp, "%s|%.2f\n", nama_pos, nominal);
+      fprintf(fp, "%s|%Lf\n", nama_pos, nominal);
       fclose(fp);
 
       printf("Data berhasil ditambahkan!\n");
